refactored Resources
